+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to the "robotcode" extension will be documented in this file
   - see [README](https://github.com/d-biehl/robotcode#using-pre-release-version)
 - rework handling VSCode test items to ensure all defined tests can be executed, also when they are ambiguous
   - see [#37](https://github.com/d-biehl/robotcode/issues/37)
+- semantic highlighting of new WHILE and EXCEPT options for RF 5.0
+-
 
 
 ##  0.7.0